Organize components, polish styling, upload photos.
