+++ client/src/components/BookingWidget.jsx
