+++ client/src/components/Header.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { Link } from "react-router-dom";
-import { UserContext } from "../UserContext";
+import { UserContext } from "./UserContext";
 
 export default function Header() {
   const {user} = useContext(UserContext);
