+++ client/src/components/PhotoGallery.jsx
