+++ client/src/pages/PlacesPage.jsx
@@ -1,8 +1,8 @@
 import { Link } from "react-router-dom";
 import { useEffect, useState } from "react";
-import AccountNav from "../AccountNav";
+import AccountNav from "../components/AccountNav";
 import axios from "axios";
-import List from "../List";
+import List from "../components/List";
 
 export default function PlacesPage() {
   const [places, setPlaces] = useState([]);
